+++ packages/adapters/ethers/src/client.ts
@@ -437,6 +437,16 @@ export class EthersAdapter {
         return await provider.request({ method: 'wallet_grantPermissions', params })
       },
 
+      revokePermissions: async session => {
+        const provider = ProviderUtil.getProvider<Provider>(CommonConstantsUtil.CHAIN.EVM)
+
+        if (!provider) {
+          throw new Error('Provider is undefined')
+        }
+
+        return await provider.request({ method: 'wallet_revokePermissions', params: [session] })
+      },
+
       sendTransaction: async data => {
         if (data.chainNamespace && data.chainNamespace !== 'eip155') {
           throw new Error(`Invalid chain namespace - Expected eip155, got ${data.chainNamespace}`)
@@ -589,6 +599,10 @@ export class EthersAdapter {
     })
   }
 
+  /**
+   * Checks the active providers and sets the provider. We call this when we initialize the adapter.
+   * @param config - The provider config
+   */
   private checkActiveProviders(config: ProviderType) {
     const walletId = SafeLocalStorage.getItem(SafeLocalStorageKeys.WALLET_ID)
     const walletName = SafeLocalStorage.getItem(SafeLocalStorageKeys.WALLET_NAME)
@@ -617,6 +631,12 @@ export class EthersAdapter {
     }
   }
 
+  /**
+   * Sets the provider and updates the local storage. We call this when we connect with external providers or via checkActiveProviders function.
+   * @param provider - The provider to set
+   * @param providerId - The provider id
+   * @param name - The name of the provider
+   */
   private async setProvider(provider: Provider, providerId: ProviderIdType, name?: string) {
     if (providerId === 'w3mAuth') {
       this.setAuthProvider()
@@ -631,9 +651,12 @@ export class EthersAdapter {
       if (provider) {
         const { addresses, chainId } = await EthersHelpersUtil.getUserInfo(provider)
         const firstAddress = addresses?.[0]
-        const caipAddress = `${this.chainNamespace}:${chainId}:${firstAddress}` as CaipAddress
+        const caipNetwork = this.caipNetworks.find(c => c.id === chainId) ?? this.caipNetworks[0]
+        const caipAddress =
+          `${this.chainNamespace}:${caipNetwork?.id}:${firstAddress}` as CaipAddress
 
-        if (firstAddress && chainId) {
+        if (firstAddress && caipNetwork) {
+          this.appKit?.setCaipNetwork(caipNetwork)
           this.appKit?.setCaipAddress(caipAddress, this.chainNamespace)
           ProviderUtil.setProviderId('eip155', providerId)
           ProviderUtil.setProvider<Provider>('eip155', provider)
@@ -871,6 +894,10 @@ export class EthersAdapter {
     }
   }
 
+  /**
+   * Syncs the account state depending on the given parameters. We call this in different conditions like when caipNetwork or caipAddress changes, when the user switches account or network.
+   * @param param0 - The address and caipNetwork. Both are optional.
+   */
   private async syncAccount({
     address,
     caipNetwork
@@ -892,9 +919,6 @@ export class EthersAdapter {
         )
 
         this.syncConnectedWalletInfo()
-        if (this.ethersConfig) {
-          this.checkActiveProviders(this.ethersConfig)
-        }
 
         if (currentCaipNetwork?.blockExplorers?.default.url) {
           this.appKit?.setAddressExplorerUrl(