+++ packages/adapters/wagmi/src/client.ts
@@ -19,12 +19,12 @@ import {
   getConnections,
   switchAccount,
   injected,
-  createConfig,
-  getConnectors
+  createConfig
 } from '@wagmi/core'
 import {
   ChainController,
   ConstantsUtil as CoreConstantsUtil,
+  CoreHelperUtil,
   StorageUtil
 } from '@reown/appkit-core'
 import type UniversalProvider from '@walletconnect/universal-provider'
@@ -188,6 +188,19 @@ export class WagmiAdapter implements ChainAdapter {
       chains: this.wagmiChains,
       transports,
       connectors
+    })
+
+    ChainController.subscribeKey('activeCaipAddress', val => {
+      const isEVMAddress = val?.startsWith('eip155:')
+      const caipNetwork = ChainController.state.activeCaipNetwork
+      const isEVMNetwork = caipNetwork?.chainNamespace === this.chainNamespace
+
+      if (caipNetwork && isEVMAddress && isEVMNetwork) {
+        this.setProfileAndBalance(
+          CoreHelperUtil.getPlainAddress(val) as Hex,
+          Number(caipNetwork.id)
+        )
+      }
     })
   }
 
@@ -494,6 +507,26 @@ export class WagmiAdapter implements ChainAdapter {
 
         return provider.request({ method: 'wallet_grantPermissions', params })
       },
+      revokePermissions: async session => {
+        if (!this.wagmiConfig) {
+          throw new Error('connectionControllerClient:revokePermissions - wagmiConfig is undefined')
+        }
+
+        const connections = getConnections(this.wagmiConfig)
+        const connection = connections[0]
+
+        if (!connection?.connector) {
+          throw new Error('connectionControllerClient:revokePermissions - connector is undefined')
+        }
+
+        const provider = (await connection.connector.getProvider()) as UniversalProvider
+
+        if (!provider) {
+          throw new Error('connectionControllerClient:revokePermissions - provider is undefined')
+        }
+
+        return provider.request({ method: 'wallet_revokePermissions', params: [session] })
+      },
 
       sendTransaction: async (data: SendTransactionArgs) => {
         if (data.chainNamespace && data.chainNamespace !== 'eip155') {
@@ -656,6 +689,10 @@ export class WagmiAdapter implements ChainAdapter {
       })
   }
 
+  private async setProfileAndBalance(address: Hex, chainId: number) {
+    await Promise.all([this.syncProfile(address, chainId), this.syncBalance(address, chainId)])
+  }
+
   private async syncAccount({
     address,
     chainId,
@@ -681,6 +718,7 @@ export class WagmiAdapter implements ChainAdapter {
       this.appKit?.resetNetwork(this.chainNamespace)
       this.appKit?.setAllAccounts([], this.chainNamespace)
       this.appKit?.setStatus(status, this.chainNamespace)
+      this.appKit?.setLoading(false)
       SafeLocalStorage.removeItem(SafeLocalStorageKeys.WALLET_ID)
       if (isAuthConnector) {
         await connector.disconnect()
@@ -717,8 +755,6 @@ export class WagmiAdapter implements ChainAdapter {
           ) {
             this.syncNetwork(address, currentChainId, true)
             await Promise.all([
-              this.syncProfile(address, currentChainId),
-              this.syncBalance(address, currentChainId),
               this.syncConnectedWalletInfo(connector),
               this.appKit?.setApprovedCaipNetworksData(this.chainNamespace)
             ])
@@ -729,11 +765,10 @@ export class WagmiAdapter implements ChainAdapter {
           const caipAddress = `eip155:${chainId}:${address}` as CaipAddress
           this.syncNetwork(address, chainId, true)
           await Promise.all([
-            this.syncProfile(address, chainId),
-            this.syncBalance(address, chainId),
             this.syncConnectedWalletInfo(connector),
             this.appKit?.setApprovedCaipNetworksData(this.chainNamespace)
           ])
+          this.appKit?.setLoading(false)
           this.appKit?.setCaipAddress(caipAddress, this.chainNamespace)
           this.appKit?.setStatus('connected', this.chainNamespace)
           // Set by authConnector.onIsConnectedHandler as we need the account type
@@ -745,21 +780,12 @@ export class WagmiAdapter implements ChainAdapter {
           }
         } else if (status === 'reconnecting') {
           this.appKit?.setLoading(true)
-          const connectors = getConnectors(this.wagmiConfig)
-          const currentConnector = connectors.find(c => c.id === connector.id)
-
-          if (currentConnector) {
-            await reconnect(this.wagmiConfig, {
-              connectors: [currentConnector]
-            })
-            this.appKit?.setLoading(false)
-          }
         }
       }
     }
   }
 
-  private async syncNetwork(address?: Hex, chainId?: number, isConnected?: boolean) {
+  private syncNetwork(address?: Hex, chainId?: number, isConnected?: boolean) {
     const caipNetwork = this.caipNetworks.find((c: CaipNetwork) => c.id === chainId)
 
     if (caipNetwork && chainId) {
@@ -774,8 +800,6 @@ export class WagmiAdapter implements ChainAdapter {
         } else {
           this.appKit?.setAddressExplorerUrl(undefined, this.chainNamespace)
         }
-
-        await this.syncBalance(address, chainId)
       }
     }
   }
@@ -1010,10 +1034,6 @@ export class WagmiAdapter implements ChainAdapter {
           this.appKit?.setCaipAddress(undefined, this.chainNamespace)
           this.appKit?.setLoading(false)
         }
-      })
-
-      provider.onIsConnected(() => {
-        provider.connect()
       })
 
       provider.onConnect(user => {