+++ packages/core/src/controllers/RouterController.ts
@@ -80,6 +80,8 @@ export interface RouterControllerState {
     | 'SwapPreview'
     | 'ConnectingMultiChain'
     | 'SwitchActiveChain'
+    | 'SmartSessionCreated'
+    | 'SmartSessionList'
   history: RouterControllerState['view'][]
   data?: {
     connector?: Connector